% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbmm.polygon.R
\name{bbmm.polygon}
\alias{bbmm.polygon}
\title{Export Routine for BBMM polygons}
\usage{
bbmm.polygon(x, crs.current, crs.utm, lev, plot = FALSE, path, indID)
}
\arguments{
\item{x, }{a data frame formatted appropriately for use in a \code{brownian.bridge} object with the requirement that spatial data columns are labeled as "Lat" and "Lon"}

\item{crs.current, }{coordinate reference system identifying the current state of the data (e.g., lat/long).  Uses standard \code{CRS} structure within quotes: CRS("+proj=longlat +zone=14 +datum=NAD83")}

\item{crs.utm, }{coordinate reference system that transforms the curent state to UTM.  Uses standard \code{CRS} structure within quotes: CRS("+proj=utm +zone=14 +datum=NAD83")}

\item{lev, }{level of the UD contour as a vector, c(50, 95) that the user is interested in.   Will work with multiple values (e.g., c(50, 95)) and will label those values
in the resultant shapefile.  Note that you don't have to put these values in as percentages (e.g., 0.50 for 50 percent).}

\item{plot, }{do you want \code{R} to plot the image on screen, default is FALSE.}

\item{path, }{file path (e.g., "C:/") using standard R path nomenclature specifying the location that the output shapefile is to be written.  path uses \code{setwd()} currently.}

\item{indID, }{Unique identifier for each individual, in quotation's, specifying the output files name which will be written to the \code{path} designated above.  Will additionally add 
a column to the attribute table with the individuals unique ID.}
}
\value{
Output is a polygon shapefile, written to the specified \code{path}, with \code{n} polygon features equivalent to the utilization distribution levels requested by the user
}
\description{
\code{bbmm.polygon} function that creates the utilization distribution contours based on the \code{bbmm.contour}
from package BBMM and exports the created contour lines as polygon shapefile for further analysis in ArcMap (or GIS program of choice).
}
\details{
This is a simple extractor function for a \code{brownian.bridge()} that exports the \code{bbmm.contour()} results from \code{R} package BBMM into a polygon shapefile.  
Since BBMM requires UTM, the functions transforms the data from the current format (defined by \code{CRS()} to UTM internally). The primary requisite is that the column names for the 
input data frame's longitude and latitude data are labeled "Lon", and "Lat", and the format needs to be appropriate for a BBMM analysis, it will error otherwise.
}
\author{
Bret A. Collier <bret@lsu.edu>
}
\references{
Horne, J. S., E. O. Garton, S. M. Krone, and J. S. Lewis.  2007.  Analyzing animal movements using Brownian bridges.  Ecology 88:2354-2363

Sawyer, H., M. J. Kauffman, R. M. Nielson and J. S. Horne.  2009.  Identifying and prioritizing ungulate migration routes for landscape-level conservation.  Ecological Applications 19: 2016-2025.
}
\keyword{moveud}

