% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveud-package.R
\docType{package}
\name{moveud-package}
\alias{moveud}
\alias{moveud-package}
\title{Simple export routines for utilization distributions for BBMM and dbbmm objects}
\description{
This package creates either a \code{SpatialLineDataFrame} or a \code{SpatialPolygonsDataFrame} and outputs individual step-specific shapefiles of contour lines
representing the utlization distribution for a specific time step for dynamic brownian bridge movement models
as detailed in Kranstauber et al. (2012).  Additionally, this package allows for exporting \code{bbmm.contour()} output to 
a polygon shapefile as well.  Note that this package uses a significant number of other R packages to stitch the various things together
so I wanted to give credit where it was due so be sure to see Description for the list of required packages I used.
}
\details{
\tabular{ll}{ Package: \tab moveud \cr Type: \tab Package\cr Version:
\tab 1.1.0\cr Date: \tab 2016-05-26\cr License: \tab GPL (>=2)\cr LazyLoad: \tab
yes\cr }
}
\author{
Bret Collier <bret@lsu.edu>
}
\references{
Kranstauber, D., R. Kays, S. D. Lapoint, M. Wikelski, and K. Safi.  2012.  A dynamic Brownian bridge movement model to estimate
utlization distributions for heterogenous animal movement.  Journal of Animal Ecology 81: 738-746. DOI: 10.1111/j.1365-2656.2012.01955.x.

Horne, J. S., E. O. Garton, S. M. Krone, and J. S. Lewis.  2007.  Analyzing animal movements using Brownian bridges.  Ecology 88:2354-2363

Sawyer, H., M. J. Kauffman, R. M. Nielson and J. S. Horne.  2009.  Identifying and prioritizing ungulate migration routes for landscape-level conservation.  
Ecological Applications 19: 2016-2025.

Byrne, M. E., J. C. McCoy, J. Hinton, M. J. Chamberlain, and B. A. Collier. 2014. Using dynamic brownian bridge movement modeling to measure temporal 
patterns of habitat selection. Journal of Animal Ecology,  83: 1234-1243.
}
\keyword{package}

