% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveud-package.R
\docType{data}
\name{rawturkey}
\alias{rawturkey}
\title{Raw turkey dataset}
\format{The format is a dataframe 
\describe{\item{Try}{Unique identifier for each GPS location attempt, NA's are allowed}
\item{FixDate}{Date for each fix taken}
\item{FixTime}{Time for each fix taken}
\item{Lat}{Latitude}
\item{Lon}{Longitude}
\item{Sats}{Number of satellites used to calculate location}
\item{HDOP}{Horizontal dilution of precision; measure of positional accuracy (lower is better)}}}
\description{
Raw turkey dataset
}
\details{
The raw turkey dataset is exactly what it says it is, a unformatted dataset which I have not turned into a \code{move} object, however, the code
for manipulation a raw dataset such as this into a \code{move} object is included below.  Additionally, the code below suggests that the user look at their
data before creating the \code{move} object, as this is a nice example of having one outlier point that causes problems when estimating the utilization
distribution.
}
\examples{

\donttest{
data(rawturkey)
head(rawturkey)
plot(rawturkey$Lon, rawturkey$Lat)
rawturkey=rawturkey[rawturkey$Lon< -98.1,]
turkey=na.omit(rawturkey)
#rawturkey$DateTime=strptime(paste(rawturkey$FixDate, rawturkey$FixTime),"\%m/\%d/\%Y \%H:\%M:\%S") 
#Create object of class: move
#move.out <- move(x=turkey$Lon, y=turkey$Lat, time=as.POSIXct(turkey$DateTime, tz="GMT"), data=turkey, proj=CRS("+proj=longlat"), animal="3057")
#out <- spTransform(move.out, CRSobj="+proj=aeqd", center=TRUE)
#turkey <- brownian.bridge.dyn(object=out, location.error=15, margin=5, window.size=21, dimSize=85)
}



}
\references{
Guthrie, J. D., M. E. Byrne, J. B. Hardin, C. O. Kochanny, K. L. Skow, R. T. Snelgrove, M. J. Butler, M. J. Peterson, M. J. Chamberlain, 
and B. A. Collier. 2011. Evaluation of a GPS backpack transmitter for wild turkey research. Journal of Wildlife Management 75:539-547.
}
\keyword{datasets}

