% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveud-package.R
\docType{data}
\name{turkey}
\alias{turkey}
\title{Turkey dataset}
\format{The format is a turkey object of Rio Grande wild turkey movements created with the R package move}
\description{
The data included in this package can be accessed using \code{data(turkey)} represent a dbbmm object for a single Rio Grande wild turkey 
tagged with a GPS transmitter in Texas during 2009.
}
\details{
This package contains 2 functions \code{move.contour} and \code{move.forud} both which create contour lines for user-specified utilization distributions for individual
time steps using the input of a \code{turkey} object from the R package move and the contour function \code{bbmm.contour} from the R package \code{BBMM}.
}
\examples{

\donttest{
data(turkey)
move.contour(turkey, 10:12, 15, c(50, 95), "C:/", "Test", crs="+proj=longlat +ellps=WGS84")
}

\donttest{
data(turkey)
move.forud(turkey, range.subset=10:12, le=6, lev=c(50, 95), ras=5, ts=.2, path="C:/", name="test", crs="+proj=longlat +ellps=WGS84", ID="123") 
}



}
\references{
Guthrie, J. D., M. E. Byrne, J. B. Hardin, C. O. Kochanny, K. L. Skow, R. T. Snelgrove, M. J. Butler, M. J. Peterson, M. J. Chamberlain, 
and B. A. Collier. 2011. Evaluation of a GPS backpack transmitter for wild turkey research. Journal of Wildlife Management 75:539-547.
}
\keyword{datasets}

